melody_engine control the logic of how/when to play/stop melody, melodies.cpp manages the add/delete/modify melodies fixed files
