+++ src/main.ino
@@ -2,7 +2,7 @@
 #include "alarm.h"
 #include "ui.h"
 #include "buttons.h"
-#include "melodies.h"
+#include "melody_engine.h"
 #include "utils.h"
 #include "config.h"
 #include <WiFi.h>
