+++ src/melodies.cpp
@@ -1,6 +1,97 @@
-#include "melodies.h"
-#include "config.h"
 
+#include "melodies.h"
+#include "Arduino.h"
+
+#define NOTE_B0    31
+#define NOTE_C1    33
+#define NOTE_CS1   35
+#define NOTE_D1    37
+#define NOTE_DS1   39
+#define NOTE_E1    41
+#define NOTE_F1    44
+#define NOTE_FS1   46
+#define NOTE_G1    49
+#define NOTE_GS1   52
+#define NOTE_A1    55
+#define NOTE_AS1   58
+#define NOTE_B1    62
+#define NOTE_C2    65
+#define NOTE_CS2   69
+#define NOTE_D2    73
+#define NOTE_DS2   78
+#define NOTE_E2    82
+#define NOTE_F2    87
+#define NOTE_FS2   93
+#define NOTE_G2    98
+#define NOTE_GS2  104
+#define NOTE_A2   110
+#define NOTE_AS2  117
+#define NOTE_B2   123
+#define NOTE_C3   131
+#define NOTE_CS3  139
+#define NOTE_D3   147
+#define NOTE_DS3  156
+#define NOTE_E3   165
+#define NOTE_F3   175
+#define NOTE_FS3  185
+#define NOTE_G3   196
+#define NOTE_GS3  208
+#define NOTE_A3   220
+#define NOTE_AS3  233
+#define NOTE_B3   247
+#define NOTE_C4   262
+#define NOTE_CS4  277
+#define NOTE_D4   294
+#define NOTE_DS4  311
+#define NOTE_E4   330
+#define NOTE_F4   349
+#define NOTE_FS4  370
+#define NOTE_G4   392
+#define NOTE_GS4  415
+#define NOTE_A4   440
+#define NOTE_AS4  466
+#define NOTE_B4   494
+#define NOTE_C5   523
+#define NOTE_CS5  554
+#define NOTE_D5   587
+#define NOTE_DS5  622
+#define NOTE_E5   659
+#define NOTE_F5   698
+#define NOTE_FS5  740
+#define NOTE_G5   784
+#define NOTE_GS5  831
+#define NOTE_A5   880
+#define NOTE_AS5  932
+#define NOTE_B5   988
+#define NOTE_C6  1047
+#define NOTE_CS6 1109
+#define NOTE_D6  1175
+#define NOTE_DS6 1245
+#define NOTE_E6  1319
+#define NOTE_F6  1397
+#define NOTE_FS6 1480
+#define NOTE_G6  1568
+#define NOTE_GS6 1661
+#define NOTE_A6  1760
+#define NOTE_AS6 1865
+#define NOTE_B6  1976
+#define NOTE_C7  2093
+#define NOTE_CS7 2217
+#define NOTE_D7  2349
+#define NOTE_DS7 2489
+#define NOTE_E7  2637
+#define NOTE_F7  2794
+#define NOTE_FS7 2960
+#define NOTE_G7  3136
+#define NOTE_GS7 3322
+#define NOTE_A7  3520
+#define NOTE_AS7 3729
+#define NOTE_B7  3951
+#define NOTE_C8  4186
+#define NOTE_CS8 4435
+#define NOTE_D8  4699
+#define NOTE_DS8 4978
+#define REST        0
 
 const int melodyWeWishYou[] = {
   NOTE_D4, 4, NOTE_G4, 4, NOTE_G4, 4, NOTE_A4, 4,
@@ -12,6 +103,7 @@ const int melodyWeWishYou[] = {
   NOTE_E4, 4, NOTE_E4, 4, NOTE_A4, 4, NOTE_FS4, 4,
   NOTE_G4, 2, NOTE_D4, 2
 };
+const int melodyWeWishYouLength = sizeof(melodyWeWishYou) / sizeof(melodyWeWishYou[0]) / 2;
 
 extern const int melodyWhiteChristmas[] = {
   NOTE_D4, 4, NOTE_G4, 4, NOTE_G4, 4, NOTE_A4, 4,
@@ -23,6 +115,7 @@ extern const int melodyWhiteChristmas[] = {
   NOTE_E4, 4, NOTE_E4, 4, NOTE_A4, 4, NOTE_FS4, 4,
   NOTE_G4, 2, NOTE_D4, 2
 };
+const int melodyWhiteChristmasLength = sizeof(melodyWhiteChristmas) / sizeof(melodyWhiteChristmas[0]) / 2;
 
 extern const int melodyJingleBells[] = {
   NOTE_E4, 4, NOTE_E4, 4, NOTE_E4, 2,
@@ -35,6 +128,7 @@ extern const int melodyJingleBells[] = {
   NOTE_E4, 4, NOTE_E4, 4, NOTE_D4, 4, NOTE_D4, 4,
   NOTE_E4, 4, NOTE_D4, 2, NOTE_G4, 2
 };
+const int melodyJingleBellsLength = sizeof(melodyJingleBells) / sizeof(melodyJingleBells[0]) / 2;
 
 extern const int melodyRudolfTheRedNosed[] = {
   NOTE_G4, 4, NOTE_E4, 4, NOTE_F4, 4, NOTE_G4, 4,
@@ -47,6 +141,7 @@ extern const int melodyRudolfTheRedNosed[] = {
   NOTE_D5, 4, NOTE_E5, 2, NOTE_F5, 4, NOTE_E5, 4,
   NOTE_D5, 4, NOTE_C5, 4, NOTE_B4, 2, NOTE_A4, 2
 };
+const int melodyRudolfTheRedNosedLength = sizeof(melodyRudolfTheRedNosed) / sizeof(melodyRudolfTheRedNosed[0]) / 2;
 
 extern const int melodySantaClausIsComin[] = {
   NOTE_G4,8,
@@ -66,6 +161,7 @@ extern const int melodySantaClausIsComin[] = {
   NOTE_D4,4, NOTE_F4,2, NOTE_D5,4,
   NOTE_C5,1
 };
+const int melodySantaClausIsCominLength = sizeof(melodySantaClausIsComin) / sizeof(melodySantaClausIsComin[0]) / 2;
 
 extern const int melodySilentNight[] = {
   NOTE_G4,-4, NOTE_A4,8, NOTE_G4,4,
@@ -92,45 +188,7 @@ extern const int melodySilentNight[] = {
   NOTE_G4,-4, NOTE_F4,8, NOTE_D4,4,
   NOTE_C4,-2
 };
-
-
-void startMelodyPreview(const int* melody, int length, int tempo, int pin) {
-  currentMelody = melody;
-  melodyLength = length;
-  melodyTempo = tempo;
-  melodyIndex = 0;
-  lastNoteTime = 0;
-  buzzerPin = pin;
-  melodyPlaying = true;
-}
-
-void updateMelodyPlayback() {
-  if (!melodyPlaying || !currentMelody) {
-    return;  // nothing to do
-  }
-
-  unsigned long now = millis();
-  if (now >= lastNoteTime) { //It advances the melody based on time
-    int note = currentMelody[melodyIndex * 2];
-    int duration = currentMelody[melodyIndex * 2 + 1];
-    int wholenote = (60000 * 4) / melodyTempo;
-    int noteDuration = (duration > 0) ? wholenote / duration : (wholenote / abs(duration)) * 1.5;
-
-    if (note > 0) {
-      tone(buzzerPin, note, noteDuration * 0.9);  // play tone for ~90% duration
-    } else {
-      noTone(buzzerPin);  // rest
-    }
-
-    lastNoteTime = now + noteDuration;
-    melodyIndex++;
-
-    if (melodyIndex >= melodyLength) {
-      melodyPlaying = false;
-      noTone(buzzerPin);  // stop at end of melody
-    }
-  }
-}
+const int melodySilentNightLength = sizeof(melodySilentNight) / sizeof(melodySilentNight[0]) / 2;
 
 
 const int* getMelodyData(int id) {
@@ -147,12 +205,12 @@ const int* getMelodyData(int id) {
 
 int getMelodyLength(int id) {
   switch (id) {
-    case 0: return sizeof(melodyWeWishYou) / sizeof(melodyWeWishYou[0]) / 2;
-    case 1: return sizeof(melodyWhiteChristmas) / sizeof(melodyWhiteChristmas[0]) / 2;
-    case 2: return sizeof(melodyJingleBells) / sizeof(melodyJingleBells[0]) / 2;
-    case 3: return sizeof(melodyRudolfTheRedNosed) / sizeof(melodyRudolfTheRedNosed[0]) / 2;
-    case 4: return sizeof(melodySantaClausIsComin) / sizeof(melodySantaClausIsComin[0]) / 2;
-    case 5: return sizeof(melodySilentNight) / sizeof(melodySilentNight[0]) / 2;
+    case 0: return melodyWeWishYouLength;
+    case 1: return melodyWhiteChristmasLength;
+    case 2: return melodyJingleBellsLength;
+    case 3: return melodyRudolfTheRedNosedLength;
+    case 4: return melodySantaClausIsCominLength;
+    case 5: return melodySilentNightLength;
     default: return 0;
   }
 }
@@ -168,11 +226,3 @@ int getMelodyTempo(int id) {
     default: return 120;
   }
 }
-
-void stopMelody() {
-  melodyPlaying = false;
-  noTone(buzzerPin);
-  Serial.println("Stop melody");
-}
-
-
